Add user role for category and comments

+++ OnlineShopApp/Controllers/CommentsController.cs
@@ -1,6 +1,9 @@
-﻿using Microsoft.AspNetCore.Mvc;
+﻿using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Identity;
+using Microsoft.AspNetCore.Mvc;
 using OnlineShopApp.Data;
 using OnlineShopApp.Models;
+using System.Data;
 
 namespace OnlineShopApp.Controllers
 {
@@ -9,11 +12,24 @@ namespace OnlineShopApp.Controllers
 
         private readonly ApplicationDbContext db;
 
-        public CommentsController(ApplicationDbContext context)
+        private readonly UserManager<ApplicationUser> _userManager;
+
+        private readonly RoleManager<IdentityRole> _roleManager;
+
+        public CommentsController(
+            ApplicationDbContext context,
+            UserManager<ApplicationUser> userManager,
+            RoleManager<IdentityRole> roleManager
+            )
         {
             db = context;
+
+            _userManager = userManager;
+
+            _roleManager = roleManager;
         }
 
+        /*
         [HttpPost]
         public IActionResult New(Comment comm)
         {
@@ -34,48 +50,84 @@ namespace OnlineShopApp.Controllers
             }
 
         }
+        */
+
+        [Authorize(Roles = "User,Editor,Admin")]
         public IActionResult Edit(int id)
         {
             try
             {
                 Comment comment = db.Comments.Find(id);
-                return View(comment);
+                if (comment.UserId == _userManager.GetUserId(User) || User.IsInRole("Admin"))
+                {
+                    return View(comment);
+                }
+                else
+                {
+                    TempData["message"] = "Nu aveți dreptul să editați comentariul";
+                    TempData["messageType"] = "alert-danger";
+                    return RedirectToAction("Index", "Products");
+                }
+                    
             }
             catch (Exception)
             {
                 TempData["message"] = $"Nu s-a putut efectua operația de editare comentariul. ID-ul {id} nu corespounde niciunui comentariu din baza de date!";
                 TempData["messageType"] = "alert-danger";
+                return RedirectToAction("Index", "Products");
             }
-            return Redirect("/Products");
         }
 
         [HttpPost]
+        [Authorize(Roles = "User,Editor,Admin")]
         public IActionResult Edit(int id, Comment requestComment)
         {
             Comment comment = db.Comments.Find(id);
-            if(ModelState.IsValid)
+
+            if (comment.UserId == _userManager.GetUserId(User) || User.IsInRole("Admin"))
             {
-                comment.Content = requestComment.Content;
-                db.SaveChanges();
-                TempData["message"] = $"Comentariul cu id-ul {id} a fost modificat cu succes!";
-                TempData["messageType"] = "alert-success";
-                return Redirect("/Products/Show/" + comment.ProductId);
+                if (ModelState.IsValid)
+                {
+                    comment.Content = requestComment.Content;
+                    db.SaveChanges();
+                    TempData["message"] = $"Comentariul cu id-ul {id} a fost modificat cu succes!";
+                    TempData["messageType"] = "alert-success";
+                    return Redirect("/Products/Show/" + comment.ProductId);
+                }
+                else
+                {
+                    return View(requestComment);
+                }
             }
-            return View(requestComment);
+            else
+            {
+                TempData["message"] = "Nu aveți dreptul să faceți modificări";
+                TempData["messageType"] = "alert-danger";
+                return RedirectToAction("Index", "Products");
+            }
         }
 
         [HttpPost]
+        [Authorize(Roles = "User,Editor,Admin")]
         public IActionResult Delete(int id)
         {
             try
             {
                 Comment comment = db.Comments.Find(id);
-                db.Comments.Remove(comment);
-                db.SaveChanges();
-                TempData["message"] = $"Comentariul cu id-ul {id} a fost șters cu succes!";
-                TempData["messageType"] = "alert-success";
-                //return Redirect("/Products/Show/" + comment.Id);
-                return Redirect("/Products/Show/" + comment.ProductId);
+                if (comment.UserId == _userManager.GetUserId(User) || User.IsInRole("Admin"))
+                {
+                    db.Comments.Remove(comment);
+                    db.SaveChanges();
+                    TempData["message"] = $"Comentariul cu id-ul {id} a fost șters cu succes!";
+                    TempData["messageType"] = "alert-success";
+                    return Redirect("/Products/Show/" + comment.ProductId);
+                }
+                else
+                {
+                    TempData["message"] = "Nu aveți dreptul să ștergeți comentariul";
+                    TempData["messageType"] = "alert-danger";
+                    return RedirectToAction("Index", "Products");
+                }
             }
 
             catch(Exception)
